feat(interface): 整合v1版本接口
